all: various minor code cleanups
Use log.Fatalf instead of utils.Fatalf, to consolidate around our own
logger.

Avoid utils.StartNode, since that sets a SIGINT handler, while we
already set our own signal handler in our main func.

Avoid setting fields to zero values in composite literals, as that's a
no-op.

Don't ignore some potentially problematic errors, such as when compiling
a regular expression, or when listing pins in ipfssync.

Deduplicate []byte conversions by using []byte globals instead of
strings.

+++ cmd/gateway/gateway.go
@@ -444,7 +444,7 @@ func main() {
 			time.Sleep(time.Second * 5)
 			h = vnode.State.Height()
 			log.Infof("[vochain info] replaying block %d at %d b/s",
-				h, int64((h-hPrev)/5))
+				h, (h-hPrev)/5)
 		}
 	}
